requirements.txt
Pillow
pillow-heif
send2trash
opencv-python
numpy
pandas
PyQt6
piexif
pymediainfo
